+++ lib/xdr/transactions/operations/claim_claimable_balance_result_code.ex
@@ -11,7 +11,8 @@ defmodule StellarBase.XDR.Operations.ClaimClaimableBalanceResultCode do
     CLAIM_CLAIMABLE_BALANCE_CANNOT_CLAIM: -2,
     CLAIM_CLAIMABLE_BALANCE_LINE_FULL: -3,
     CLAIM_CLAIMABLE_BALANCE_NO_TRUST: -4,
-    CLAIM_CLAIMABLE_BALANCE_NOT_AUTHORIZED: -5
+    CLAIM_CLAIMABLE_BALANCE_NOT_AUTHORIZED: -5,
+    CLAIM_CLAIMABLE_BALANCE_TRUSTLINE_FROZEN: -6
   ]
 
   @enum_spec %XDR.Enum{declarations: @declarations, identifier: nil}